Make LoopVectorization functionality opt-in using package extensions
