+++ Project.toml
@@ -1,6 +1,6 @@
 name = "ImageMorphology"
 uuid = "787d08f9-d448-5407-9aad-5290dd7ab264"
-version = "0.4.7"
+version = "0.5.0"
 
 [deps]
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
@@ -10,6 +10,14 @@ LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 TiledIteration = "06e1c1a7-607b-532d-9fad-de7d9aa2abac"
+
+[weakdeps]
+LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
+ImageMetadata = "bc367c6b-8a6b-528e-b4bd-a4b897500b49"
+
+[extensions]
+LoopVectorizationExt = "LoopVectorization"
+ImageMetadataExt = "ImageMetadata"
 
 [compat]
 DataStructures = "0.19"
@@ -24,9 +32,10 @@ julia = "1.6"
 [extras]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ImageMetadata = "bc367c6b-8a6b-528e-b4bd-a4b897500b49"
+LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Documenter", "ImageMetadata", "OffsetArrays", "Suppressor", "Test"]
+test = ["Documenter", "ImageMetadata", "LoopVectorization", "OffsetArrays", "Suppressor", "Test"]
