+++ test/extreme_filter.jl
@@ -97,30 +97,32 @@
     for (se_gen_func, se_name) in Any[(strel_diamond, "diamond"), (strel_box, "box")]
         test_name = "optimization: $se_name"
         @testset "$test_name" begin
-            # ensure the optimized implementation work equivalently to the generic fallback implementation
-            for T in Any[Bool, Int, N0f8, Gray{N0f8}, Gray{Float64}, Float64]
-                for N in (1, 2, 3)
-                    sz = N == 3 ? (32, 32, 5) : ntuple(_ -> 32, N)
-                    img = T == Int ? rand(1:10, sz...) : rand(T, sz...)
-                    for r in (1, 3)
-                        dims_list = ntuple(i -> ntuple(identity, i), N)
-                        for dims in dims_list
-                            se = se_gen_func(ntuple(_ -> 2r + 1, N), dims)
-                            ref = ImageMorphology._extreme_filter_generic!(max, similar(img), img, se)
-                            out = extreme_filter(max, img, se)
-                            @test out == ref
+            for accelerate_mode ∈ (true, false)
+                # ensure the optimized implementation work equivalently to the generic fallback implementation
+                for T in Any[Bool, Int, N0f8, Gray{N0f8}, Gray{Float64}, Float64]
+                    for N in (1, 2, 3)
+                        sz = N == 3 ? (32, 32, 5) : ntuple(_ -> 32, N)
+                        img = T == Int ? rand(1:10, sz...) : rand(T, sz...)
+                        for r in (1, 3)
+                            dims_list = ntuple(i -> ntuple(identity, i), N)
+                            for dims in dims_list
+                                se = se_gen_func(ntuple(_ -> 2r + 1, N), dims)
+                                ref = ImageMorphology._extreme_filter_generic!(max, similar(img), img, se)
+                                out = extreme_filter(max, img, se, accelerate_mode)
+                                @test out == ref
 
-                            # `maybe_floattype` is used to pre-allocate the output for
-                            # higher level operators (e.g., tophat) to avoid overflow
-                            # behavior. We need to ensure it works correctly.
-                            # https://github.com/JuliaImages/ImageMorphology.jl/issues/104
-                            out = similar(img, ImageMorphology.maybe_floattype(T))
-                            extreme_filter!(max, out, img, se)
-                            @test out == ref
+                                # `maybe_floattype` is used to pre-allocate the output for
+                                # higher level operators (e.g., tophat) to avoid overflow
+                                # behavior. We need to ensure it works correctly.
+                                # https://github.com/JuliaImages/ImageMorphology.jl/issues/104
+                                out = similar(img, ImageMorphology.maybe_floattype(T))
+                                extreme_filter!(max, out, img, se, accelerate_mode)
+                                @test out == ref
 
-                            imgc = centered(img)
-                            outc = extreme_filter(max, imgc, se)
-                            @test outc == centered(out)
+                                imgc = centered(img)
+                                outc = extreme_filter(max, imgc, se, accelerate_mode)
+                                @test outc == centered(out)
+                            end
                         end
                     end
                 end
